Updated banner text styling to match the home page; added text hover effect; created attractions.css
Made changes based on @KimberlyO367 's updates, I like your lightsaber so I added some glowing effect! The font family "'Pacifico', cursive" looks like this on my laptop, it's not as the same as the pic in the slack channel I think?

(not sure why the previous one was automatically approved, so posting again)